started constructing the frontend
game mode will be chosen in a panel beneath the game footer. this currently has 3 tabs at the top for each game mode that open up a component. these components will soon be the selected game mode options. also did a bit of restructuring of the components dir. now there are folders for tiles, buttons and controls, each with an index.ts so that they can all be accessed easily.